requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
click==8.0.3
colorama==0.4.4
daphne==3.0.2
decouple==0.0.7
dj-database-url==2.1.0
Django==3.2.8
django-cors-headers==4.0.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
email-validator==1.1.3
gunicorn==21.2.0
idna==2.10
jsonschema==3.2.0
Pillow==8.3.2
psycopg2-binary==2.9.3
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.25.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
websocket-client==0.59.0
whitenoise==6.5.0
urllib3==1.26.7
